feat : Register / RegisterDetail 구조 변경 (Discussion #22 참고)
