+++ ottogether/src/pages/Register/RegisterDetail.module.css
@@ -163,86 +163,6 @@
 	}
 }
 
-/* agreement fieldset */
-.agreement-section {
-	p {
-		/* margin-bottom: 0.2rem; */
-		line-height: 1.2;
-		color: var(--color-grey-300);
-	}
-
-	ol {
-		margin-bottom: 1rem;
-		padding: 1.5rem 2rem;
-		border: 1px solid var(--foreground);
-		list-style-type: decimal;
-		color: var(--color-grey-100);
-		display: flex;
-		flex-direction: column;
-		gap: 1rem;
-	}
-
-	label[for="agree"],
-	label[for="disagree"] {
-		display: flex;
-		align-items: center;
-		gap: 1rem;
-		cursor: pointer;
-	}
-
-	fieldset {
-		border: none;
-		padding: 0;
-		display: flex;
-		flex-direction: column;
-		gap: 0.8rem;
-
-		legend {
-			margin-bottom: 1rem;
-		}
-	}
-
-	input[type="radio"] {
-		appearance: none;
-		margin-right: 1rem;
-		width: 1.2rem;
-		height: 1.2rem;
-		border-radius: 50%;
-		border: 2px solid var(--color-grey-300);
-		background-color: var(--foreground);
-		position: relative;
-		cursor: pointer;
-
-		display: flex;
-		justify-content: center;
-		align-items: center;
-
-		transition: background-color 0.2s ease, border-color 0.2s ease;
-	}
-
-	input[type="radio"]:checked::before {
-		content: "✔︎";
-		font-size: 0.6rem;
-		color: var(--foreground);
-
-		position: absolute;
-		display: flex;
-		justify-content: center;
-		align-items: center;
-
-		width: 100%;
-		height: 100%;
-		border-radius: 50%;
-		border: 2px solid var(--color-primary-300);
-		background-color: var(--primary);
-	}
-
-	input[type="radio"]:focus {
-		outline: 2px solid var(--color-error-300);
-		outline-offset: 2px;
-	}
-}
-
 /* button */
 .register-button {
 	font-family: "Poppins";
